Shuffle include path ordering
This re-orders the include paths provided to protoc during a schema
compile, to have the user-provided paths be passed in first before the
WDLC compiler installed include path. This allows the user to pass in
paths to files that may over-ride the installed set that come with the
compiler (e.g test traits).