Remove tc_e field from test traits c and d
Following a code review, we could not definitively answer why this
field was added.  Its presence, complicates maintenance at this point
in time: generated code for those traits now breaks tests in TestTDM
in openweave-core.  The expedient workaround is to remove the field
for now; we can re-add it with a suitable test case in openweave-core.