Pip installation from local site packages
Added options to the Python installation scripts to allow pointing to a
local directory that contains staged site local packages for pip to
pick-up during virtual env setup.

This leverages the --find-links and --no-index options to pick up
packags from a local directory.

This is wired to configure through the LOCAL_SITE_PACKAGES_DIR
environment variable, allowing users to set that variable to point to
the directory of choice.