Use pip to setup VirtualEnv
This replaces the use of easy_install to setup virtualenv with pip since
the former seems to give Github the fits with more recent versions of
virtualenv - It starts failing installing some of the dependent packages of virtualenv
(zipp, see https://github.com/LibreTime/libretime/issues/952). Even putting that
workaround in isn't sufficient, as other failures ensue. Using pip seems to be
the more modern way of setting up virtualenv, so follow the masses if it ain't causing issues.