Fix Instance ID Validator
This fixes the instance ID validator to permit instance IDs for traits
in resources to start at a number other than 0.

Caononically, a trait instance is usually numbered from 0 - however in
cases where there are derived traits involved, while each of the derived
traits may be different in trait, given they share the same base trait,
they each have to have a different instance ID to permit polymorphic
access to those base traits.

E.g Trait A is extended by B and C (i.e B->A, C->A).

When instantiating both B and C in a resource, they each have to have
different instance IDs.

This allows a down-casting of each to of type A and be accessed
independently.